+++ R/ggwithinstats.R
@@ -161,7 +161,7 @@ ggwithinstats <- function(data,
   if (isTRUE(results.subtitle)) {
     # preparing the bayes factor message
     if (type == "parametric" && isTRUE(bf.message)) {
-      caption <-
+      caption_df <-
         function_switch(
           test = test,
           # arguments relevant for expression helper functions
@@ -172,13 +172,14 @@ ggwithinstats <- function(data,
           bf.prior = bf.prior,
           top.text = caption,
           paired = TRUE,
-          output = "caption",
           k = k
         )
+
+      caption <- caption_df$expression[[1]]
     }
 
     # extracting the subtitle using the switch function
-    subtitle <-
+    subtitle_df <-
       function_switch(
         test = test,
         # arguments relevant for expression helper functions
@@ -195,6 +196,8 @@ ggwithinstats <- function(data,
         conf.level = conf.level,
         k = k
       )
+
+    subtitle <- subtitle_df$expression[[1]]
   }
 
   # return early if anything other than plot